% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxymax.R
\name{read_oxymax}
\alias{read_oxymax}
\title{Read in OxyMax csv data file.}
\usage{
read_oxymax(file)
}
\arguments{
\item{file}{OxyMax csv file.}
}
\value{
A data.frame.
}
\description{
OxyMax Respirometer machine collects data on gas concentrations in an
enclosed environment. The output from the machine is a csv file with standard
}
\examples{

\dontrun{
fi <- system.file("extdata", "oxymax_data.csv", package = "readmisc")
read_oxymax(fi)
}

}
